Use type cast to avoid conversation error with C99 standard.
JerryScript-DCO-1.0-Signed-off-by: Robert Sipka rsipka.uszeged@partner.samsung.com

+++ jerry-core/ecma/operations/ecma-objects.cpp
@@ -615,7 +615,7 @@ ecma_op_object_get_property_names (ecma_object_t *obj_p, /**< object */
       ecma_string_t *name_p = ecma_get_string_from_value (*iter.current_value_p);
 
       lit_string_hash_t hash = name_p->hash;
-      uint32_t bitmap_row = hash / bitmap_row_size;
+      uint32_t bitmap_row = (uint32_t) (hash / bitmap_row_size);
       uint32_t bitmap_column = hash % bitmap_row_size;
 
       if ((own_names_hashes_bitmap[bitmap_row] & (1u << bitmap_column)) == 0)
@@ -647,7 +647,7 @@ ecma_op_object_get_property_names (ecma_object_t *obj_p, /**< object */
         if (!(is_enumerable_only && !ecma_is_property_enumerable (prop_iter_p)))
         {
           lit_string_hash_t hash = name_p->hash;
-          uint32_t bitmap_row = hash / bitmap_row_size;
+          uint32_t bitmap_row = (uint32_t) (hash / bitmap_row_size);
           uint32_t bitmap_column = hash % bitmap_row_size;
 
           bool is_add = true;
@@ -686,7 +686,7 @@ ecma_op_object_get_property_names (ecma_object_t *obj_p, /**< object */
                                             true);
 
           lit_string_hash_t hash = name_p->hash;
-          uint32_t bitmap_row = hash / bitmap_row_size;
+          uint32_t bitmap_row = (uint32_t) (hash / bitmap_row_size);
           uint32_t bitmap_column = hash % bitmap_row_size;
 
           if ((names_hashes_bitmap[bitmap_row] & (1u << bitmap_column)) == 0)
@@ -804,7 +804,7 @@ ecma_op_object_get_property_names (ecma_object_t *obj_p, /**< object */
       ecma_string_t *name_p = names_p[i];
 
       lit_string_hash_t hash = name_p->hash;
-      uint32_t bitmap_row = hash / bitmap_row_size;
+      uint32_t bitmap_row = (uint32_t) (hash / bitmap_row_size);
       uint32_t bitmap_column = hash % bitmap_row_size;
 
       if ((names_hashes_bitmap[bitmap_row] & (1u << bitmap_column)) == 0)
